templates/*.j2: Add the templates/*.j2 option
It seems to be ops_template.py doesn't chnage the path
of the CWD, which is different from the lookup() method
behavior.  Anyway, this change adds the additional
templates/* files for all the j2 include statement
to make it work both cases.  We might remove those in
the future, once we have a consistent behavior but
for now, we just add the 'templates/*.j2' option.

Signed-off-by: Kei Nohguchi <kei@nohguchi.com>